refactor: Move AppConfigForTestContainers into TestContainerLayers
